feat(JS lib): add a client JS lib for the Communicator
similar to trame-client js-api, trame-iframe js-api provides an agnostic library
that can be installed in a regular JS project.
trame-iframe js-api could be installed in a host application (could be
react/streamlit/vue/... based) to proxy the trame API calls through the
iframe to the Communicator and the trame application.
The goal is to replicate the trame-client js-api features to make it
transparent to the consumer code.
API:
- trame.state.set
- trame.state.update
- trame.state.watch
- trame.trigger

This is still possible to directly interact with the Communicator from
an host application by manually sending/listening to events.